refactor git-path tests
